Well a lot of changes
Rocketcar is in the midst of a lot of changes. The main reason for this
commit is to assure that the remote git builds on a clean setup. The
reason for not committing more frequently is to minimize adding
temporary resources.

I've updated the rube parser, implemented a car picker, fixed several
issues. Also some work has been put into the the level editor project.